requirements.txt
fastapi>=0.116.0
uvicorn>=0.34.0
python-multipart>=0.0.20
pydantic>=2.11.0
PyYAML>=6.0.2
openai>=1.58.1
python-docx>=1.1.2
pypdf>=5.1.0
openpyxl>=3.1.5
reportlab>=4.2.5
pytest>=8.3.4
httpx>=0.28.1
